Isolate: convert to TypedData
This is the least straightforward of the TypedData conversions
due to a possible bugfix in rb_context_create_isolate_value.

I believe the original call to Data_Wrap_Struct from
rb_context_create_isolate_value intentionally omitted the dmark
callback as a possible optimization.  It's no longer possible to
have a one-off definition with TypedData.  I also suspect there
were cases where the original code would be unsafe if the
original Isolate Ruby object died before the Context.

In any case, the performance cost of an extra GC mark call is
probably negligible, here; and the safest course of action is
to always mark a data type consistently.