replace (int)RSTRING_LEN(...) with RSTRING_LENINT(...)
Casting the result of RSTRING_LEN on 64-bit long systems with 32-bit
ints can lead to truncation errors when dealing with gigantic strings.
RSTRING_LENINT was introduced in Ruby 1.9.2 to avoid this problem, so
make use of it.  It's used by by several standard extensions (openssl,
zlib) and expected to remain part of the supported Ruby C API.