avoid malloc in single-ref case of free_context
Triggering an allocation inside a `free' callback doesn't help
in low memory situations; so operate directly on the ContextInfo
being freed when possible instead of creating a short-lived copy
of it.

For the `isolate_info->refs() > 1' case, we'll leak slightly
less memory by freeing the short-lived allocation in case
pthread_create(3) fails.