Fix segfault, handle symbol conversion correctly
Symbol::Description() normally returns a string, but when the symbol
does not have a description, it returns undefined.

The String::Utf8Value constructor requires that a v8::Context is active
for the ToString operation it executes. ToString is a no-op for strings
(hence no crash) but not for undefined.

Add a Context::Scope to fix that.

Fixes: https://github.com/rubyjs/mini_racer/issues/318