Handle two-byte string alignment padding
For compatibility with old clients of the serializer wire format, V8
inserts a padding byte when a two-byte string is not 16 bits aligned,
so skip those.

We don't have to emit padding ourselves, V8 handles unaligned strings
just fine.

Messages can have an arbitrary amount of padding at the end. Skip that
as well.

No test because hitting the right conditions in V8 is rather difficult
and brittle.